Added pojo/refactoring
