+++ service/src/main/java/uk/gov/hmcts/reform/fpl/enums/ccd/fixedlists/InterimEndDateType.java
@@ -1,0 +1,6 @@
+package uk.gov.hmcts.reform.fpl.enums.ccd.fixedlists;
+
+public enum InterimEndDateType {
+    END_OF_PROCEEDINGS,
+    NAMED_DATE
+}
